requirements.txt
streamlit
torch
modelscope
transformers
Pillow
pandas
openpyxl
flash-attn
accelerate
sentencepiece
